LayoutTests/platform/ios/fast/canvas/offscreen-giant-transfer-to-imagebitmap-expected.txt

CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 67108864).
CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 67108864).
CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 67108864).
CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 67108864).
CONSOLE MESSAGE: Canvas area exceeds the maximum limit (width * height > 67108864).
Testing height: 100 type: none
Got exception: "InvalidStateError: The object is in an invalid state."
Testing height: 100 type: 2d
Got context: "[object OffscreenCanvasRenderingContext2D]"
Got result with size: 100x100
Testing height: 100 type: webgl
Got context: "[object WebGLRenderingContext]"
Got result with size: 100x100
Testing height: 100 type: webgl2
Got context: "[object WebGL2RenderingContext]"
Got result with size: 100x100
Testing height: 100 type: webgpu
Got context: "[object GPUCanvasContext]"
Got result with size: 100x100
Testing height: 100 type: bitmaprenderer
Got context: "[object ImageBitmapRenderingContext]"
Got result with size: 100x100
Testing height: 100000000 type: none
Got exception: "InvalidStateError: The object is in an invalid state."
Testing height: 100000000 type: 2d
Got context: "[object OffscreenCanvasRenderingContext2D]"
Got exception: "UnknownError: The operation failed for an unknown transient reason (e.g. out of memory)."
Testing height: 100000000 type: webgl
Got context: "[object WebGLRenderingContext]"
Got exception: "UnknownError: The operation failed for an unknown transient reason (e.g. out of memory)."
Testing height: 100000000 type: webgl2
Got context: "[object WebGL2RenderingContext]"
Got exception: "UnknownError: The operation failed for an unknown transient reason (e.g. out of memory)."
Testing height: 100000000 type: webgpu
Got context: "[object GPUCanvasContext]"
Got exception: "UnknownError: The operation failed for an unknown transient reason (e.g. out of memory)."
Testing height: 100000000 type: bitmaprenderer
Got context: "[object ImageBitmapRenderingContext]"
Got exception: "UnknownError: The operation failed for an unknown transient reason (e.g. out of memory)."
PASS successfullyParsed is true

TEST COMPLETE
